pyproject.toml
[project]
name = "resume-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf2>=2.8.7",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.5",
]
